Change window UID to Int64 with persistent counter in config.json
- WindowTracker._nextUid is now long (Int64), monotonically increasing
- Counter persisted in %APPDATA%\WindowMover\config.json across restarts
- LoadNextUid() reads from config on startup; SaveNextUid() writes after each assignment
- Eliminates UID collisions: counter never resets, even across WindowMover restarts
- WindowSnapshot.Uid updated to long
- AppConfig class added for extensible JSON config storage

+++ src/WindowMover.Core/Models/WindowSnapshot.cs
@@ -8,9 +8,9 @@ public class WindowSnapshot
 {
     /// <summary>
     /// Unique identifier assigned via SetProp during the current session.
-    /// Only meaningful within the session that created it; ignored after reboot.
+    /// Monotonically increasing Int64, persisted in config.json across app restarts.
     /// </summary>
-    public int Uid { get; set; }
+    public long Uid { get; set; }
 
     /// <summary>
     /// Process name (e.g., "firefox", "devenv").
